refactor: integrate secretvaults into nilrag
