+++ .env.example
@@ -1,0 +1,21 @@
+# Your Credentials:
+NILLION_ORG_DID="Add your ORG DID here"
+NILLION_ORG_SECRET_KEY="Add you SECRET_KEY here"
+
+# Update this with your own schema ID after creating a new schema
+SCHEMA_ID=
+CLUSTERS_SCHEMA_ID=
+QUERY_ID=
+
+# Cluster Config:
+# nildb-node 1:
+URL1='https://nildb-nx8v.nillion.network'
+DID1='did:nil:testnet:nillion1qfrl8nje3nvwh6cryj63mz2y6gsdptvn07nx8v'
+
+# nildb-node 2:
+URL2='https://nildb-p3mx.nillion.network'
+DID2='did:nil:testnet:nillion1uak7fgsp69kzfhdd6lfqv69fnzh3lprg2mp3mx'
+
+# nildb-node 3:
+URL3='https://nildb-rugk.nillion.network'
+DID3='did:nil:testnet:nillion1kfremrp2mryxrynx66etjl8s7wazxc3rssrugk'
