+++ pyproject.toml
@@ -13,9 +13,10 @@ dependencies = [
     "aiohttp>=3.11.14",
     "ecdsa>=0.19.0",
     "faker>=37.0.0",
-    "nilql==0.0.0a12",
+    "nilql==0.0.0a13",
     "numpy<2.0.0",
     "PyJWT[crypto]~=2.10.1",
+    "secretvaults==0.0.0a10",
     "sentence-transformers>=3.3.1",
 ]
